% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EAS.R
\name{add.zeros}
\alias{add.zeros}
\title{add zero abundances if missing}
\usage{
add.zeros(
  data = NULL,
  ID.year = "ID.year",
  species = "species",
  abundance = "abundance",
  ID.year.dep = NULL,
  species.dep = NULL
)
}
\arguments{
\item{data}{dataframe including "abundance" per "ID.year" for all "species" (= raw data)}

\item{ID.year}{columname of ID.year in \code{data}}

\item{species}{columname of species names in \code{data}}

\item{abundance}{columname of abundance in \code{data}}

\item{ID.year.dep}{columns in \code{data} which values depend on "ID.year"
(values will be duplicated for added zero abundances)}

\item{species.dep}{columns in \code{data} which values depend on "species"
(values will be duplicated for added zero abundances)}
}
\value{
\code{add.zeros()} returns the original dataframe extended by the newly added zero abundances.
}
\description{
This function adds zero abundances for each species-ID.year combination which
is not present in the given dataframe
}
