% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EAS.R
\name{mod.conv}
\alias{mod.conv}
\title{Check model convergence (Rhat, NEff, ...)}
\usage{
mod.conv(
  model.list = NULL,
  model.name = NULL,
  td = NULL,
  plot.conv = FALSE,
  spec = NULL
)
}
\arguments{
\item{model.list}{list with different models (of one species) for comparisons}

\item{model.name}{names of models (e.g. family and zero-inflated coefficients).
If \code{NULL} model.names will be the names of model.list elements}

\item{td}{the maximum treedepth used in the given models}

\item{plot.conv}{if \code{TRUE}, graphical output is given}

\item{spec}{name of species (for title of plot)}
}
\value{
\code{mod.conv()} returns a dataframe including model convergence parameters:
maximum Rhat, minimum Neff/Ntot, maximum MCSE/SD,
the number of divergent transitions and maximum treedepth.

The function optionally returns a graphical output if \code{plot.stats = TRUE}.
}
\description{
This function extracts parameters of model convergence per model coefficient
to compare the given models of one species in terms of model performance, e.g.,
\itemize{
\item maximum Rhat: this should be close to 1.00 and ideally not exceed 1.02
\item min.Neff/Ntot: the minimum effective sample size divided by the total sample size.
The higher the portion (ideally close to 1) the better.
\item maximum MCSE/SD: ideally, the Monte Carlo standard error (MCSE) shouldn't
exceed 2\% of the standard deviation (SD)
\item divergent transitions: divergent transitions are often an indicator of
a bad model fit and should be avoided
\item maximum treedepth: this should not exceed the treedepth defined in the models
}

Note that additional checks of model performance are necessary,
e.g. by checking trace plots and doing posterior prediction checks.
}
