% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EAS.R
\name{mod.stat}
\alias{mod.stat}
\title{Check model statistics (propZ, mean, ...)}
\usage{
mod.stat(
  model.list = NULL,
  model.name = NULL,
  response = NULL,
  plot.stats = FALSE,
  spec = NULL
)
}
\arguments{
\item{model.list}{list with different models (of one species) for comparisons}

\item{model.name}{names of models (e.g. family and zero-inflated coefficients)}

\item{response}{name of response variable (character)}

\item{plot.stats}{if \code{TRUE}, graphical output is given}

\item{spec}{name of species (for title of plot)}
}
\value{
\code{mod.stat()} returns a dataframe including the above mentioned model statistics
for the proportion of zeros, minimum, maximum, mean, median, and SD of the
response variable for all given models.

The function optionally returns a graphical output inlcuding the observed value
of the raw data as well as median with 50\% and 95\% CrI of the simulated values
if \code{plot.stats = TRUE}.
}
\description{
This function extracts model statistics (different quantiles of simulated values,
observed value, BayesP-value) for the proportion of zeros (propZ), minimum, maximum,
mean, median, and SD of the response variable for different models of one species
to simplify model selection
}
