% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/norms_coef_nacc.R
\docType{data}
\name{norms_coef_nacc}
\alias{norms_coef_nacc}
\title{Regression Coefficients from NACC normative data}
\format{
a tibble(dataframe)
}
\usage{
data(norms_coef_nacc)
}
\description{
Linear regression results of UDSNB 3.0 cognitive tests from NACC normative data.
The data consisted of baseline administration of the UDSNB 3.0 for 5031 NACC participants, who had a global Clinical Dementia Rating Scale (CDR) of 0(“normal cognition)” and were aged 70 years or older.
}
\examples{
data(norms_coef_nacc)
head(norms_coef_nacc)
}
\keyword{datasets}
