% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EAS.R
\name{obs.eff}
\alias{obs.eff}
\title{observer effects based on total abundance}
\usage{
obs.eff(
  data = NULL,
  ID = "ID",
  abundance = "abundance",
  year = "year",
  OE = 0.25,
  add.df = TRUE
)
}
\arguments{
\item{data}{dataframe including "abundance" per site "ID" and survey "year"
for all species (= raw data)}

\item{ID}{columname of site ID in \code{data}}

\item{abundance}{columname of abundance in \code{data}}

\item{year}{columname of year in \code{data}}

\item{OE}{threshold for classifying observer effects, should be a value >0
and <1.}

\item{add.df}{if \code{TRUE}, returns a dataframe with observer effect per site and
year}
}
\value{
\code{obs.eff()} returns a dataframe including the implemented raw data
(in \code{data}) as well as an additional column for observer effect (\code{OE_xx})
per site and year and the default column \code{abundance}.

The function optionally returns the dataframe \code{df.oe} containing observer
effects per site and year if \code{add.df = TRUE}.
This dataframe has the following columns:
\itemize{
\item \code{ID}:            site ID
\item \code{year}:          year of survey
\item \code{Ab.total}:      total abundance per site and year (summed up across all
species)
\item \code{mean.Ab.total}: mean total abundance (\code{Ab.total}) per site
\item \code{Ab.prop}:       proportional abundance per site and year
(= \code{Ab.total} / \code{mean.Ab.total} per site)
\item \code{OE_xx}:         observer effect levels "negative", "none", and "positive"
per site and year. "xx" is the used threshold in \%.
}
}
\description{
This function calculates observer effects based on the proportional total
abundance (= total abundance per site and year / mean total abundance per site).
For each site and year, observer effects are classified in three classes
using a defined threshold
\itemize{
\item if proportional abundance < (1-threshold), the effect is classified as
"negative".
\item if proportional abundance > (1+threshold), the effect is classified as
"positive".
\item if (1-threshold) <= proportional abundance <= (1+threshold), the effect is
classified as "none".
}
}
