% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample_dat.R
\docType{data}
\name{sample_dat}
\alias{sample_dat}
\title{Sample Data for z-score Calculation}
\format{
a tibble(dataframe)
}
\usage{
data(sample_dat)
}
\description{
A sample dataset with the demographic variables and UDSNB 3.0 cognitive test variables for z-score calculation.
You should prepare your dataset to make sure it is in the standard format and had the correct variable names.
}
\examples{
data(sample_dat)
head(sample_dat)
str(sample_dat)
}
\keyword{datasets}
