% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/norms.R
\name{uds_z}
\alias{uds_z}
\title{Calculate the z-scores and impairment indicators of UDS3 cognitive tests using EAS or NACC norms}
\usage{
uds_z(dat, tests, norms = "eas", impair_sd = 1, verbose = T)
}
\arguments{
\item{dat}{a dataframe or tibble, contains columns of demographic covariates and UDS3 cognitive tests}

\item{tests}{character or a character vector, the name of the cognitive test(s)}

\item{norms}{specify if you want to use EAS norms(norms='eas') or NACC norms(norms='nacc'), default is 'eas'}

\item{impair_sd}{a positive number, how many SD worse than the mean should be defined as impairment,}

\item{verbose}{logical, verbose =T adds z-scores and indicators of impairment,}
}
\value{
a tibble(dataframe)
}
\description{
This function generates z-scores and impairment indicators of UDS3 cognitive tests. Users can secify whether to use EAS or NACC norms.
}
