% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OEFS_Paper_Functions.R
\name{weight}
\alias{weight}
\title{Weights based on habitat representativity}
\usage{
weight(
  df.habitat = NULL,
  habitat = "habitat",
  area = "area",
  df.data = NULL,
  ID = "ID",
  year = "year",
  by_spec = FALSE,
  species = "species",
  by_reg = FALSE,
  region = "region",
  add.df = TRUE
)
}
\arguments{
\item{df.habitat}{dataframe including present "habitats" and the total "area"
of the habitat of the studied region (this is the target share)}

\item{habitat}{column name of habitat in \code{df.habitat} and \code{df.data}}

\item{area}{column name of area (share) in \code{df.habitat}}

\item{df.data}{dataframe including site "ID" for each "year" the site was
surveyed and the assigned "habitat". You can use your raw data per species
since the function condenses the df to its unique site-year combination.}

\item{ID}{column name of site ID in \code{df.data}}

\item{year}{column name of year in \code{df.data}}

\item{by_spec}{logical indicating whether weights should be calculated per
species (\code{by_spec = TRUE}, e.g., if observations were excluded species-specific
resulting in different annual site-subsets per species)}

\item{species}{column name of species name in \code{df.data}
(only needed if \code{by_spec = TRUE})}

\item{add.df}{if \code{TRUE}, returns a dataframe with weights per habitat and
year (and species)}
}
\value{
\code{weight()} returns a dataframe including the implemented raw data \code{df.data}
as well as \code{weight} per habitat and year (and species if \code{by_spec = TRUE})
and the default columns \code{habitat}, \code{ID}, \code{year} (and \code{species}).
If \code{by_spec = FALSE}, the weights per habitat and year are the same for all
species.

The function optionally returns the dataframe \code{df.w} containing
weights per habitat and year (and species if \code{by_spec = TRUE}).
This dataframe has the following columns:
\itemize{
\item \code{year}: year of survey
\item \code{habitat}: habitat
\item \code{N.sites}: number of surveyed sites per habitat and year (and species)
\item \code{actual}: actual annual share of the habitat: number of surveyed sites per
habitat and year (and species)/total number of surveyed sites per year
(and species)
\item \code{target}: target share of the habitat
\item \code{weight}: weight for each habitat-year (or habitat-year-species) combination
\item (\code{species}: species name (only if \code{by_spec = TRUE}))
}
}
\description{
This function calculates weights based on annual habitat representativity:
\itemize{
\item if a habitat was underrepresented in a given year,
the annual weight will be >1
\item if a habitat was overrepresented  in a given year,
the annual weight will be <1
\item if a habitat was represented according to its share in a given year,
the annual weight will be 1
}
}
