% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/norms_coef_eas.R
\docType{data}
\name{norms_coef_eas}
\alias{norms_coef_eas}
\title{Regression Coefficients from EAS norms data}
\format{
a tibble(dataframe)
}
\usage{
data(norms_coef_eas)
}
\description{
Linear regression results of UDSNB 3.0 cognitive tests from EAS normative data
}
\examples{
data(norms_coef_eas)
head(norms_coef_eas)
}
\keyword{datasets}
