% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/norms.R
\name{uds_z}
\alias{uds_z}
\title{Calculate the z-scores and impairment indicators of UDS3 cognitive tests}
\usage{
uds_z(dat, tests, norms = "eas", impair_sd = 1, out_mean_sd = F)
}
\arguments{
\item{dat}{a dataframe or tibble which contains columns of demographic covariates and UDS3 cognitive tests in numeric format.

       It should have the four demographic variables named as \code{female}(binary, 1 indicates female), \code{age}(continuous),
       \code{educyrs}(continuous) and \code{black_race}(binary, 1 indicates non-hispanic black).
       Also, it should have at least one of the following 17 UDSNB 3.0 cognitive test variables(also see Details):
       \code{mocascore},\code{verbatimi},\code{paraphrasei},\code{verbatimd},\code{paraphrased},\code{bensonscorei},\code{bensonscored},\code{numspancorf},
       \code{numspancorb},\code{minttotal},\code{fwords60sec},\code{lwords60sec},\code{flword},\code{animals60sec},\code{vegetables60sec},\code{tr_a1},\code{tr_b1}.

       All the demographic variables and cognitive test variables need to be numeric.}

\item{tests}{character or a character vector, the name of the cognitive test(s)}

\item{norms}{specify if you want to use EAS norms(norms='eas') or NACC norms(norms='nacc'), default is 'eas'}

\item{impair_sd}{a positive number, how many SD worse than the mean should be defined as impairment, default is 1.}

\item{out_mean_sd}{logical, out_mean_sd=T additionally adds mean and sd estimates for each cognitive test in the output dataframe, default is F}
}
\value{
a tibble(dataframe) with the demographically adjusted z-score and impairment indicator
       for each of the specified cognitive tests.
}
\description{
\code{uds_z} generates demographically adjusted z-scores and impairment indicators of UDS3 cognitive tests. Users can specify whether
to use EAS or NACC norms.
}
\details{
Below is the list of UDSNB 3.0 tests, the corresponding variables and item numbers on the UDSNB 3.0 form.\cr
- Montreal Cognitive Assessment(MoCA): \code{mocascore}(1f)\cr
- Craft Story 21 Recall(Immediate): \code{verbatimi}(3a), \code{paraphrasei}(3b)\cr
- Benson Complex Figure Copy: \code{bensonscorei}(4a) \cr
- Number Span Test-Forward: \code{numspancorf}(5a) \cr
- Number Span Test(Backward):\code{numspancorb}(6a) \cr
- Category Fluency:\code{animals60sec}(7a), \code{vegetables60sec}(7b) \cr
- Trail Making Test: \code{tr_a1}(8a), \code{tr_b1}(8b) \cr
- Craft Story 21 Recall(Delayed): \code{verbatimd}(9a), \code{paraphrased}(9b) \cr
- Benson Complex Figure Recall: \code{bensonscored}(10a) \cr
- Multilingual Naming Test (MINT): \code{minttotal}(11a) \cr
- Verbal Fluency(Phonemic Test): \code{fwords60sec}(12a), \code{lwords60sec}(12d), \code{flword}(12g)

Regression estimates from normative data are used to generate gender, age, education, black race adjusted z-scores.
Use \code{?norms_coef_eas} and \code{?norms_coef_nacc} to see the documentation of regression estimates from the two norms.

The mean is calculated as:
\deqn{\hat{Y}=\hat{b0}+ \hat{b1}*female + \hat{b_2}*(age-77) + \hat{b3}*(educyrs-16)+ \hat{b4}*black}
For trail A1 and trail B1, as a higher score indicates a worse performance, z-score is calculated as
\deqn{-(Y - \hat{Y})/\hat{\sigma}}
For all the other cognitive tests where a higher score indicates a better performance, z-score is calculated as
\deqn{(Y - \hat{Y})/\hat{\sigma}}
Thus, an impairment can be defined as a z-score below a certain cutoff.
For example, using a cutoff of -1 means that a subject is defined as impaired on this test if their performance is 1SD worse than the population mean.
The indicator of impairment is generated as:
 \deqn{impair =1 if z-score<=-1}
You can specify how many SD worse than the population mean should be defined as impairment by providing \code{impair_sd}(default is 1).
1SD and 1.5SD are some common choices.
}
\examples{
# here is a sample datset used to calculate z-scores. You should prepare your dataset in this standard format.
head(sample_dat)
# calculate z-scores and the impairment indicators for tr_a1 and tr_b1 with NACC norms and 1.5 SD to define impairment.
uds_z(sample_dat, c('tr_a1','tr_b1'), norms = 'nacc', impair_sd = 1.5)
# calculate the z-score and the impaiment indicator for minttotal with EAS norms and 1SD to define impairment
# Also output mean and sd estimates in addition to the z-scores and the impairment indicators.
uds_z(sample_dat, 'minttotal', norms = 'eas', impair_sd = 1, out_mean_sd = T)
}
\seealso{
\url{https://github.com/JiyueQin/eas}
}
\author{
Jiyue Qin
}
